+++ client/acp/src/model/sections.js
@@ -1,9 +1,9 @@
-export const SECTION_ACTIVITY = 'sectionActivity';
-
 export const SECTION_AWARDS = 'sectionAwards';
 
 export const SECTION_GRANTING = 'sectionGranting';
 
 export const SECTION_MANAGEMENT = 'sectionManagement';
+
+export const SECTION_REWARDS = 'sectionRewards';
 
 export const SECTION_SETTINGS = 'sectionSettings';