+++ client/acp/src/view/widget/tab-manager.js
@@ -4,22 +4,22 @@ import React from 'react';
 import {connect} from 'react-redux';
 
 import {setSection} from '../../action/actions';
-import ActivityView from './activity-view';
 import AwardsListView from './awards-list-view';
 import GrantingView from './granting-view';
 import Management from './management';
+import Rewards from './rewards';
 import * as Sections from '../../model/sections';
 import {getSection, getSections} from '../../model/selector/selectors';
 
 class TabManager extends React.Component {
     createTabContent(section) {
         switch (section) {
-            case Sections.SECTION_ACTIVITY:
-                return <ActivityView/>;
             case Sections.SECTION_AWARDS:
                 return <AwardsListView />;
             case Sections.SECTION_GRANTING:
                 return <GrantingView/>;
+            case Sections.SECTION_REWARDS:
+                return <Rewards/>;
             case Sections.SECTION_MANAGEMENT:
                 return <Management/>;
             case Sections.SECTION_SETTINGS: