+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Add an edit mode for multiple awards
 - Add support for plugin driven image upload
-- Add activity section to ACP
+- Add Rewards section to ACP
 - Add protection for an image upload service
 - Change UX for the creation of awards
 - Change UX for Overview section